backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
yfinance>=0.2.31
pandas>=2.1.3
numpy>=1.26.0
requests>=2.31.0
pydantic>=2.5.0
scikit-learn>=1.3.2
tenacity>=8.2.3

# Technical Analysis
mplfinance>=0.12.10b0
textblob>=0.17.1
beautifulsoup4>=4.12.2
newsapi-python>=0.2.7

# Caching & Performance
redis>=5.0.1
fastapi-cache2>=0.2.1
structlog>=23.2.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
pytest-cov>=4.1.0
locust>=2.19.1

# Development
flake8>=6.1.0
black>=23.11.0
mypy>=1.7.0
isort>=5.12.0
